requirements.txt
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
streamlit==1.25.0
plotly==5.15.0
textblob==0.17.1
nltk==3.8.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.1
pytest==7.4.0
black==23.7.0
flake8==6.1.0
